Use strtol to take user input.
Now we can have years up to Gregorian 9223372036854774641
(Erisian INT64_MAX (9223372036854775807)

+++ src/convert.h
@@ -16,4 +16,4 @@ ddate_greg_to_eris(struct ddate *dd, int32_t year, int32_t day);
 	Like greg_to_eris, but takes a month and day of month argument
 */
 ddate_error
-ddate_greg_ymd_to_eris(struct ddate *dd, int32_t year, int32_t month, int32_t day);
+ddate_greg_ymd_to_eris(struct ddate *dd, int64_t year, int64_t month, int64_t day);
